pyproject.toml
[project]
name = "ogs-buddy"
version = "0.1.0"
description = "OGS game analysis and review support"
authors = [
    {name = "Terrence Brannon", email = "terrence.brannon@gmail.com"},
]
dependencies = [
    "Selenium-Screenshot>=1.7.0",
    "auto-selenium>=1.0.1",
    "loguru>=0.6.0",
    "PySimpleGUI>=4.60.3",
    "traitlets>=4.3.3",
]
readme = "README.md"
license = {text = "MIT"}

requires-python = ">=3.6.*"



[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
